Factor method for run_summary from dataset.py, into tf_kernel/helpers.py
